Create figureOutAllVariousOutcomes
Now comes the fun part! We need to create a function. It will take two parameters (ie. the two choices made) and then return the winning choice.

When programming a game like this, you have to first figure out all the various outcomes. One outcome is that the choice the user makes is equal to the choice the computer makes.
Why do we call the parameters choice1 and choice2 instead of making use of the variable names userChoice and computerChoice?

Because parameter names are meant to be general. This makes them more flexible so that later we can pass any variables or values through the function. In our case, we will pass computerChoice and userChoice to the function.

For example, what if this was a game involving two human choices (ie. not a human against the computer). It wouldn't make sense to have computerChoice as a parameter!